codes/python/requirements.txt
async-timeout==4.0.3
bluepy==1.3.0
buildhat==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorzero==2.0
gpiozero==2.0.1
idna==3.10
paho-mqtt==2.1.0
pyserial==3.5
python-vlc==3.0.21203
redis==5.2.0
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.3